Fixed RFPhaseNoise (for nonlinear RF)
* When using nonlinear RF, RFPhaseSpread now works as inteded
* For linear RF, scaling is currently incorrect

+++ inc/IO/ProgramOptions.hpp
@@ -253,7 +253,7 @@ private: // simulation parameters
     double meshshiftx;
     double meshshifty;
     uint32_t steps_per_Ts;
-    uint32_t steps_per_Trev;
+    double steps_per_Trev;
     int32_t renormalize;
     double rotations;
     uint32_t rotationtype;
